+++ src/dispatching/dialogue/get_chat_id.rs
@@ -1,5 +1,4 @@
-use crate::types::CallbackQuery;
-use teloxide_core::types::{ChatId, Message};
+use crate::types::{CallbackQuery, ChatId, Message, Update};
 
 /// Something that may has a chat ID.
 pub trait GetChatId {
@@ -16,5 +15,11 @@ impl GetChatId for Message {
 impl GetChatId for CallbackQuery {
     fn chat_id(&self) -> Option<ChatId> {
         self.message.as_ref().map(|mes| mes.chat.id)
+    }
+}
+
+impl GetChatId for Update {
+    fn chat_id(&self) -> Option<ChatId> {
+        self.chat().map(|chat| chat.id)
     }
 }